+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A way to configure inverted pins (#912)
 - Added API to check a GPIO-pin's interrupt status bit (#929)
 - A `embedded_io_async::Read` implementation for `UsbSerialJtag` (#889)
+- `RtcClock::get_xtal_freq`, `RtcClock::get_slow_freq` (#957)
 
 ### Changed